Implement the Creator and ConcreteCreator classes.
Add a static method in Rental to prevent instantiation without the required validations.

Implement a method for converting strings to uppercase.

Verify that the car IDs from the Car and Rental classes are identical.

Implement a facade to simplify the main function.

+++ ACar.h
@@ -7,7 +7,7 @@ class ACar
 protected:
 	short _id;
 	std::string _model;
-	short _pricePerDay;
+	float _pricePerDay;
 	CarStatus _status;
 public:
 	ACar(short id, const std::string& model, float pricePerDay, CarStatus status);
